pyproject.toml
[tool.poetry]
name = "habits-tracker"
version = "0.1.0"
description = ""
authors = ["Dmitry Minin <d.n.minin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.2.4"
dotenv = "^0.9.9"
psycopg2-binary = "^2.9.10"
ipython = "^9.4.0"
djangorestframework = "^3.16.1"
django-filter = "^25.1"
djangorestframework-simplejwt = "^5.5.1"
celery = "^5.5.3"
redis = "^6.4.0"
django-celery-beat = "^2.8.1"
coverage = "^7.10.2"
requests = "^2.32.4"
drf-yasg = "^1.21.10"
django-cors-headers = "^4.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
